refactor: use-timeout으로 파일 이름 변경 (#447)
